refactor functions (diff and derivative)
